requirements.txt
# Core Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Google Gemini API
google-generativeai>=0.3.2

# Data Processing (using compatible versions with pre-built wheels)
pandas>=2.0.0
numpy>=1.24.0

# RAG & Embeddings
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Voice Processing
SpeechRecognition>=3.10.0
gTTS>=2.5.0
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
